+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add cache height stats.
+
 ### Changed
 
 - Disable Bloom filters on database to diagnose memory leak.